+++ lazycompass-cli/src/commands/mod.rs
@@ -1,6 +1,7 @@
 mod agg;
 mod config;
 mod database;
+mod indexes;
 mod init;
 mod insert;
 mod query;
@@ -9,6 +10,7 @@ mod upgrade;
 
 pub(crate) use agg::run_agg;
 pub(crate) use config::run_config;
+pub(crate) use indexes::run_indexes;
 pub(crate) use init::run_init;
 pub(crate) use insert::run_insert;
 pub(crate) use query::run_query;